+++ src/App.js
@@ -1,12 +1,11 @@
 import React from "react";
 import { BrowserRouter as Switch, Route } from "react-router-dom";
 
-import HomePage from './pages/HomePage'
-import AboutPage from './pages/AboutPage'
+import HomePage from "./pages/HomePage";
+import AboutPage from "./pages/AboutPage";
 import FundingPage from "./pages/FundingPage";
-
+import FundRequestDetailPage from "../src/pages/FundRequestDetailPage/FunRequestDetailPage";
 import Navbar from "./components/Navbar";
-import FunDetailRequestInfo from "./components/FunDetailRequestInfo/FunDetailRequestInfo";
 
 function App() {
   return (
@@ -16,7 +15,11 @@ function App() {
         <Route path="/" exact component={HomePage} />
         <Route path="/about" exact component={AboutPage} />
         <Route path="/funding" exact component={FundingPage} />
-        <Route path="/funDetailPage" exact component={FunDetailRequestInfo} />
+        <Route
+          path="/fundRequestDetailPage"
+          exact
+          component={FundRequestDetailPage}
+        />
       </Switch>
     </div>
   );
