--- src/pages/FunDetailRequestPage/FunDetailQuestPage.js
@@ -1,12 +1,0 @@
-import React from "react";
-import FunDetailRequestInfo from "../../components/FunDetailRequestInfo/FunDetailRequestInfo";
-
-function FunDetailQuestPage() {
-  return (
-    <div>
-      <FunDetailRequestInfo />
-    </div>
-  );
-}
-
-export default FunDetailQuestPage;
